fs/f2fs: Improve resolution of the 4.9.42 merge
This is the diff between commit a2f057876a9c ("FROMLIST: f2fs: sanity
check checkpoint segno and blkoff") and commit 0f442c5b2e4a ("f2fs:
sanity check checkpoint segno and blkoff"), which are ultimately the
same commit but the latter is the one taken in the stable tree. This
diff avoids a merge conflict in 4.9.144.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>